+++ forui/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.13.0 (Next)
+## 0.13.0
 This update focuses on polishing & improving the usability of existing widgets.
 
 ### Animations
@@ -78,6 +78,16 @@ FFCheckbox(
 * Add `FBottomNavigationBarStyle.backgroundFilter`.
 
 
+### `FBreadcrumb`
+We've added support for an alternative popover menu which uses `FItem` and more closely resembles Shadcn/ui's popover
+menu suited for desktop.
+
+* Add `FBreadcrumbItem.collapsedTiles(...)`.
+
+* **Breaking** Change `FBreadcrumbItem.collapsed` to use `FItem` instead of `FTile` - use `FBreadcrumb.collapsedTiles(...)`
+  instead.
+
+
 ### `FButton`
 * Add `FButton.actions`.
 * Add `FButton.shortcuts`.
@@ -141,6 +151,10 @@ to build more complex widgets.
 
 
 ### `FPopoverMenu`
+We've added support for an alternative popover menu which uses `FItem` and more closely resembles Shadcn/ui's popover
+menu suited for desktop.
+
+* Add `FPopoverMenu.tiles(...)`
 * Add `FPopoverMenu.barrierSemanticsLabel`.
 * Add `FPopoverMenu.barrierSemanticsDismissible`.
 * Add `FPopoverMenu.builder`.
@@ -149,6 +163,7 @@ to build more complex widgets.
 * Change `FPopoverMenu`'s animation to be origin aware.
 * Change `FPopoverMenu(...)`'s `controller` to be optional.
 * **Breaking** Remove `FPopoverMenu.automatic` - This was a bad abstraction in hindsight, use `FPopoverMenu.new` instead.
+* **Breaking** Change `FPopoverMenu(...)` to use `FItem`s instead of `FTile`s - use `FPopoverMenu.tiles(...)` instead.
 
 
 ### `FPortal`